cf crawler, contest feature for geeksforgeeks atcoder  added

+++ projects/cf_users_problem_solving/contests/codeforces.py
@@ -17,7 +17,7 @@ def codeforces(futuTime):
 			if i['startTimeSeconds'] <= tomo_time:
 				secrof['title'] = i['name']
 				secrof['startTime'] = i['startTimeSeconds']
-				secrof['duration'] = i['durationSeconds']/60
+				secrof['duration'] = int(i['durationSeconds']/60)
 				secrof['link'] = 'https://codeforces.com/contests'
 
 				forces.append(secrof)
